Refactor DataTable to extract SortableHead and update styles
Extracted `SortableHead` component into its own file for better reusability and maintainability. Updated consistent class naming conventions with `core-` prefix across components. Improved Storybook stories to align with standard patterns and added additional configuration options.
