+++ packages/core-ui/src/components/ui/data-table/primitives/index.ts
@@ -3,4 +3,5 @@ export * from './debaunced-input';
 export * from './filter';
 export * from './fuzzy-filter';
 export * from './pagination';
+export * from './sortable-head';
 export * from './view-options';
